requirements.txt
yfinance>=0.2.37
pandas>=2.0.0
numpy>=1.24.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
fredapi>=0.5.1
beautifulsoup4>=4.12.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.28.0
sse-starlette>=2.0.0
sentence-transformers>=3.0.0
